Updates after some more tests.

+++ src/main/java/com/coolcat/Operatable.java
@@ -1,5 +1,5 @@
 package com.coolcat;
 
 public interface Operatable {
-    Integer operate(Node node);
+    Double operate(Node node);
 }
